🚚 Rename Pydantic v1 file for new include
