+++ tests/test_tutorial/test_schema_extra_example/test_tutorial001_py310_pv1.py
@@ -6,7 +6,7 @@ from ...utils import needs_py310, needs_pydanticv1
 
 @pytest.fixture(name="client")
 def get_client():
-    from docs_src.schema_extra_example.tutorial001_py310_pv1 import app
+    from docs_src.schema_extra_example.tutorial001_pv1_py310 import app
 
     client = TestClient(app)
     return client
